+++ integration/e2e/images/images.go
@@ -9,6 +9,7 @@ var (
 	Memcached        = "memcached:1.6.1"
 	Minio            = "minio/minio:RELEASE.2019-12-30T05-45-39Z"
 	Consul           = "consul:0.9"
+	ETCD             = "gcr.io/etcd-development/etcd:v3.4.7"
 	DynamoDB         = "amazon/dynamodb-local:1.11.477"
 	BigtableEmulator = "shopify/bigtable-emulator:0.1.0"
 	Cassandra        = "rinscy/cassandra:3.11.0"