+++ pkg/cortex/cortex.go
@@ -8,11 +8,6 @@ import (
 	"net/http"
 	"os"
 
-	"google.golang.org/grpc/health/grpc_health_v1"
-
-	"github.com/cortexproject/cortex/pkg/configs"
-	"github.com/cortexproject/cortex/pkg/storegateway"
-
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/pkg/errors"
@@ -20,9 +15,11 @@ import (
 	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/server"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/health/grpc_health_v1"
 	"gopkg.in/yaml.v2"
 
 	"github.com/cortexproject/cortex/pkg/alertmanager"
+	"github.com/cortexproject/cortex/pkg/api"
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/chunk/cache"
 	"github.com/cortexproject/cortex/pkg/chunk/encoding"
@@ -30,7 +27,8 @@ import (
 	"github.com/cortexproject/cortex/pkg/chunk/storage"
 	chunk_util "github.com/cortexproject/cortex/pkg/chunk/util"
 	"github.com/cortexproject/cortex/pkg/compactor"
-	"github.com/cortexproject/cortex/pkg/configs/api"
+	"github.com/cortexproject/cortex/pkg/configs"
+	configAPI "github.com/cortexproject/cortex/pkg/configs/api"
 	"github.com/cortexproject/cortex/pkg/configs/db"
 	"github.com/cortexproject/cortex/pkg/distributor"
 	"github.com/cortexproject/cortex/pkg/flusher"
@@ -43,7 +41,9 @@ import (
 	"github.com/cortexproject/cortex/pkg/ring/kv/memberlist"
 	"github.com/cortexproject/cortex/pkg/ruler"
 	"github.com/cortexproject/cortex/pkg/storage/tsdb"
+	"github.com/cortexproject/cortex/pkg/storegateway"
 	"github.com/cortexproject/cortex/pkg/util"
+	"github.com/cortexproject/cortex/pkg/util/grpc/healthcheck"
 	"github.com/cortexproject/cortex/pkg/util/runtimeconfig"
 	"github.com/cortexproject/cortex/pkg/util/services"
 	"github.com/cortexproject/cortex/pkg/util/validation"
@@ -73,6 +73,7 @@ type Config struct {
 	PrintConfig bool       `yaml:"-"`
 	HTTPPrefix  string     `yaml:"http_prefix"`
 
+	API              api.Config               `yaml:"api"`
 	Server           server.Config            `yaml:"server"`
 	Distributor      distributor.Config       `yaml:"distributor"`
 	Querier          querier.Config           `yaml:"querier"`
@@ -111,6 +112,7 @@ func (c *Config) RegisterFlags(f *flag.FlagSet) {
 	f.BoolVar(&c.PrintConfig, "print.config", false, "Print the config and exit.")
 	f.StringVar(&c.HTTPPrefix, "http.prefix", "/api/prom", "HTTP path prefix for Cortex API.")
 
+	c.API.RegisterFlags(f)
 	c.Server.RegisterFlags(f)
 	c.Distributor.RegisterFlags(f)
 	c.Querier.RegisterFlags(f)
@@ -177,12 +179,12 @@ func (c *Config) Validate(log log.Logger) error {
 
 // Cortex is the root datastructure for Cortex.
 type Cortex struct {
-	target             ModuleName
-	httpAuthMiddleware middleware.Interface
+	target ModuleName
 
 	// set during initialization
 	serviceMap map[ModuleName]services.Service
 
+	api           *api.API
 	server        *server.Server
 	ring          *ring.Ring
 	overrides     *validation.Overrides
@@ -198,7 +200,7 @@ type Cortex struct {
 	dataPurger    *purger.DataPurger
 
 	ruler        *ruler.Ruler
-	configAPI    *api.API
+	configAPI    *configAPI.API
 	configDB     db.DB
 	alertmanager *alertmanager.MultitenantAlertmanager
 	compactor    *compactor.Compactor
@@ -231,7 +233,8 @@ func New(cfg Config) (*Cortex, error) {
 	}
 
 	cortex.serviceMap = serviceMap
-	cortex.server.HTTP.Handle("/services", http.HandlerFunc(cortex.servicesHandler))
+	cortex.api.RegisterServiceMapHandler(http.HandlerFunc(cortex.servicesHandler))
+
 	return cortex, nil
 }
 
@@ -255,7 +258,6 @@ func (t *Cortex) setupAuthMiddleware(cfg *Config) {
 				}
 			},
 		}
-		t.httpAuthMiddleware = middleware.AuthenticateUser
 	} else {
 		cfg.Server.GRPCMiddleware = []grpc.UnaryServerInterceptor{
 			fakeGRPCAuthUniaryMiddleware,
@@ -263,7 +265,7 @@ func (t *Cortex) setupAuthMiddleware(cfg *Config) {
 		cfg.Server.GRPCStreamMiddleware = []grpc.StreamServerInterceptor{
 			fakeGRPCAuthStreamMiddleware,
 		}
-		t.httpAuthMiddleware = fakeHTTPAuthMiddleware
+		cfg.API.HTTPAuthMiddleware = fakeHTTPAuthMiddleware
 	}
 }
 
@@ -321,7 +323,7 @@ func (t *Cortex) Run() error {
 	// before starting servers, register /ready handler and gRPC health check service.
 	// It should reflect entire Cortex.
 	t.server.HTTP.Path("/ready").Handler(t.readyHandler(sm))
-	grpc_health_v1.RegisterHealthServer(t.server.GRPC, newHealthCheck(sm))
+	grpc_health_v1.RegisterHealthServer(t.server.GRPC, healthcheck.New(sm))
 
 	// Let's listen for events from this manager, and log them.
 	healthy := func() { level.Info(util.Logger).Log("msg", "Cortex started") }