+++ pkg/querier/frontend/queue_test.go
@@ -2,6 +2,7 @@ package frontend
 
 import (
 	"context"
+	"strconv"
 	"testing"
 
 	"github.com/go-kit/kit/log"
@@ -11,13 +12,16 @@ import (
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
 
-func setupFrontend(t *testing.T, config Config) *Frontend {
+func setupFrontend(config Config) (*Frontend, error) {
 	logger := log.NewNopLogger()
 
 	frontend, err := New(config, logger, nil)
-	require.NoError(t, err)
+	if err != nil {
+		return nil, err
+	}
+
 	defer frontend.Close()
-	return frontend
+	return frontend, nil
 }
 
 func testReq(ctx context.Context) *request {
@@ -35,7 +39,8 @@ func TestDequeuesExpiredRequests(t *testing.T) {
 	userID := "1"
 	userID2 := "2"
 
-	f := setupFrontend(t, config)
+	f, err := setupFrontend(config)
+	require.NoError(t, err)
 
 	ctx := user.InjectOrgID(context.Background(), userID)
 	expired, cancel := context.WithCancel(ctx)
@@ -83,4 +88,46 @@ func TestDequeuesExpiredRequests(t *testing.T) {
 	}
 	_, ok = f.queues[userID2]
 	require.Equal(t, false, ok)
+}
+
+func BenchmarkGetNextRequest(b *testing.B) {
+	var config Config
+	flagext.DefaultValues(&config)
+	config.MaxOutstandingPerTenant = 2
+
+	const numTenants = 50
+
+	frontends := make([]*Frontend, 0, b.N)
+
+	for n := 0; n < b.N; n++ {
+		f, err := setupFrontend(config)
+		if err != nil {
+			b.Fatal(err)
+		}
+
+		for i := 0; i < config.MaxOutstandingPerTenant; i++ {
+			for j := 0; j < numTenants; j++ {
+				userID := strconv.Itoa(j)
+				ctx := user.InjectOrgID(context.Background(), userID)
+
+				err = f.queueRequest(ctx, testReq(ctx))
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		}
+
+		frontends = append(frontends, f)
+	}
+
+	ctx := context.Background()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		for j := 0; j < config.MaxOutstandingPerTenant*numTenants; j++ {
+			_, err := frontends[i].getNextRequest(ctx)
+			if err != nil {
+				b.Fatal(err)
+			}
+		}
+	}
 }