+++ pkg/storegateway/metadata_fetcher_metrics.go
@@ -1,4 +1,4 @@
-package querier
+package storegateway
 
 import (
 	"sync"
@@ -10,7 +10,7 @@ import (
 
 // This struct aggregates metrics exported by Thanos MetaFetcher
 // and re-exports those aggregates as Cortex metrics.
-type metaFetcherMetrics struct {
+type MetadataFetcherMetrics struct {
 	// Maps userID -> registry
 	regsMu sync.Mutex
 	regs   map[string]*prometheus.Registry
@@ -26,8 +26,8 @@ type metaFetcherMetrics struct {
 	// blocks_meta_modified
 }
 
-func newMetaFetcherMetrics() *metaFetcherMetrics {
-	return &metaFetcherMetrics{
+func NewMetadataFetcherMetrics() *MetadataFetcherMetrics {
+	return &MetadataFetcherMetrics{
 		regs: map[string]*prometheus.Registry{},
 
 		syncs: prometheus.NewDesc(
@@ -53,13 +53,13 @@ func newMetaFetcherMetrics() *metaFetcherMetrics {
 	}
 }
 
-func (m *metaFetcherMetrics) addUserRegistry(user string, reg *prometheus.Registry) {
+func (m *MetadataFetcherMetrics) AddUserRegistry(user string, reg *prometheus.Registry) {
 	m.regsMu.Lock()
 	m.regs[user] = reg
 	m.regsMu.Unlock()
 }
 
-func (m *metaFetcherMetrics) registries() map[string]*prometheus.Registry {
+func (m *MetadataFetcherMetrics) registries() map[string]*prometheus.Registry {
 	regs := map[string]*prometheus.Registry{}
 
 	m.regsMu.Lock()
@@ -71,7 +71,7 @@ func (m *metaFetcherMetrics) registries() map[string]*prometheus.Registry {
 	return regs
 }
 
-func (m *metaFetcherMetrics) Describe(out chan<- *prometheus.Desc) {
+func (m *MetadataFetcherMetrics) Describe(out chan<- *prometheus.Desc) {
 
 	out <- m.syncs
 	out <- m.syncFailures
@@ -80,7 +80,7 @@ func (m *metaFetcherMetrics) Describe(out chan<- *prometheus.Desc) {
 	out <- m.synced
 }
 
-func (m *metaFetcherMetrics) Collect(out chan<- prometheus.Metric) {
+func (m *MetadataFetcherMetrics) Collect(out chan<- prometheus.Metric) {
 	data := util.BuildMetricFamiliesPerUserFromUserRegistries(m.registries())
 
 	data.SendSumOfCounters(out, m.syncs, "blocks_meta_syncs_total")